requirements.txt
python-multipart
fastapi
uvicorn[standard]
torch
torchvision
torchaudio
transformers
Pillow